pyproject.toml
[tool.poetry]
name = "pytranscribe"
version = "0.1.0"
description = "Python Module to Convert AWS Transcribe Data to Transcription Format"
authors = ["David Weinberg <david.weinberg@anlayzere.com>"]
license = "Apache2.0"
readme = "README.md"

packages = [
    { include = "./PyTranscribe" },
]

[tool.poetry.dependencies]
python = "^3.10"
jupyterlab = "^4.0.8"
jupyterlab-quarto = "^0.2.8"
nbdev = "^2.3.13"
boto3 = "^1.35.14"
filetype = "^1.2.0"
pathlib = "^1.0.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
